move course hooks to other file
